refactor (tests): added test db engine and update tests imports accordingly
