+++ agenta-backend/agenta_backend/tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 import asyncio
 
-from agenta_backend.models.db_engine import DBEngine
+from agenta_backend.models.db_engine.engines import test_db_engine as db_engine
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -15,12 +15,8 @@ def event_loop():
     asyncio.set_event_loop(res)
     res._close = res.close  # type: ignore
 
-    # Initialize database and create tables
-    db_engine = DBEngine()
-    res.run_until_complete(db_engine.init_db())
-
     yield res
 
     res.run_until_complete(db_engine.remove_db())  # drop database
-    res.run_until_complete(db_engine.close())  # close connections to database
+    res.run_until_complete(db_engine.close_db())  # close connections to database
     res._close()  # close event loop # type: ignore
