+++ agenta-backend/agenta_backend/tests/variants_main_router/test_variant_evaluators_router.py
@@ -3,7 +3,7 @@ import httpx
 import pytest
 import asyncio
 
-from agenta_backend.models.db_engine import db_engine
+from agenta_backend.models.db_engine.shared import db_engine
 from agenta_backend.models.api.evaluation_model import EvaluationStatusEnum
 from agenta_backend.models.db_models import (
     AppDB,
