+++ agenta-backend/agenta_backend/tests/variants_user_profile_router/test_user_profile.py
@@ -5,7 +5,7 @@ import pytest
 from sqlalchemy.future import select
 
 from agenta_backend.models.db_models import UserDB
-from agenta_backend.models.db_engine import db_engine
+from agenta_backend.models.db_engine.shared import db_engine
 from agenta_backend.models.api.user_models import User
